Validation in progress

+++ internal/data/daily_goals.go
@@ -11,6 +11,7 @@ import (
 // represents a goals entry in the sytem
 type Goals struct {
 	Goal_id      int64     `json:"goal_id"`
+	User_id      int64     `json:"user_id"`
 	Goal_text    string    `json:"goal_text"`
 	Is_completed bool      `json:"is_completed"`
 	Target_date  time.Time `json:"target_date"`
@@ -32,9 +33,9 @@ type GoalsModel struct {
 // Adds new todo entry into the database
 func (m *GoalsModel) Insert(goals *Goals) error {
 	query := `
-		INSERT INTO daily_goals (goal_text, is_completed, target_date)
-		VALUES ($1, $2, $3)
-		RETURNING goal_id, created_at`
+        INSERT INTO daily_goals (user_id, goal_text, is_completed, target_date)
+        VALUES ($1, $2, $3, $4)
+        RETURNING goal_id, created_at`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
@@ -42,6 +43,7 @@ func (m *GoalsModel) Insert(goals *Goals) error {
 	return m.DB.QueryRowContext(
 		ctx,
 		query,
+		goals.User_id,
 		goals.Goal_text,
 		goals.Is_completed,
 		goals.Target_date,
@@ -49,16 +51,17 @@ func (m *GoalsModel) Insert(goals *Goals) error {
 }
 
 // Retrieve list of all daily goal entries from the database
-func (m *GoalsModel) GoalList() ([]*Goals, error) {
+func (m *GoalsModel) GoalList(userID int64) ([]*Goals, error) {
 	query := `
-        SELECT goal_id, goal_text, target_date, is_completed
+        SELECT goal_id, user_id, goal_text, target_date, is_completed, created_at
         FROM daily_goals
+        WHERE user_id = $1
         ORDER BY created_at DESC`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	rows, err := m.DB.QueryContext(ctx, query)
+	rows, err := m.DB.QueryContext(ctx, query, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -68,7 +71,7 @@ func (m *GoalsModel) GoalList() ([]*Goals, error) {
 
 	for rows.Next() {
 		g := &Goals{}
-		err := rows.Scan(&g.Goal_id, &g.Goal_text, &g.Target_date, &g.Is_completed)
+		err := rows.Scan(&g.Goal_id, &g.User_id, &g.Goal_text, &g.Target_date, &g.Is_completed, &g.Created_at)
 		if err != nil {
 			return nil, err
 		}
@@ -83,25 +86,41 @@ func (m *GoalsModel) GoalList() ([]*Goals, error) {
 }
 
 // DeleteGoal removes a goal entry from the database using its ID
-func (m *GoalsModel) DeleteGoal(goalID int64) error {
+func (m *GoalsModel) DeleteGoal(goalID int64, userID int64) error {
 	query := `
-	DELETE FROM daily_goals WHERE goal_id = $1`
+	DELETE FROM daily_goals WHERE goal_id = $1 and user_id = $2`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	_, err := m.DB.ExecContext(ctx, query, goalID)
-	return err
+	result, err := m.DB.ExecContext(ctx, query, goalID, userID)
+	if err != nil {
+		return err
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rowsAffected == 0 {
+		return sql.ErrNoRows
+	}
+
+	return nil
 }
 
 // Get the goal info based on the goal
 func (m *GoalsModel) GetGoalByID(id int64) (*Goals, error) {
 	stmt := `
-	SELECT goal_id, goal_text, is_completed, target_date FROM daily_goals WHERE goal_id = $1`
+    SELECT goal_id, user_id, goal_text, is_completed, target_date, created_at 
+    FROM daily_goals 
+    WHERE goal_id = $1`
+
 	row := m.DB.QueryRow(stmt, id)
 
 	var g Goals
-	err := row.Scan(&g.Goal_id, &g.Goal_text, &g.Is_completed, &g.Target_date)
+	err := row.Scan(&g.Goal_id, &g.User_id, &g.Goal_text, &g.Is_completed, &g.Target_date, &g.Created_at)
 	if err != nil {
 		return nil, err
 	}
