+++ internal/data/sessions.go
@@ -11,6 +11,7 @@ import (
 // represents a session entry in the sytem
 type Sessions struct {
 	Session_id   int64     `json:"session_id"`
+	User_id      int64     `json:"user_id"`
 	Title        string    `json:"title"`
 	Description  string    `json:"description"`
 	Subject      string    `json:"subject"`
@@ -28,6 +29,8 @@ func ValidateSessions(v *validator.Validator, sessions *Sessions) {
 	v.Check(validator.MaxLength(sessions.Description, 50), "description", "must not be more than 50 bytes long")
 	v.Check(validator.NotBlank(sessions.Subject), "subject", "This field cannot be left blank")
 	v.Check(validator.MaxLength(sessions.Subject, 50), "subject", "must not be more than 50 bytes long")
+	v.Check(validator.IsValidDate(sessions.Start_date), "start_date", "Start date must be provided")
+	v.Check(validator.IsValidDate(sessions.End_date), "end_date", "End date must be provided")
 }
 
 type SessionsModel struct {
@@ -37,9 +40,9 @@ type SessionsModel struct {
 // Adds new todo entry into the database
 func (m *SessionsModel) Insert(sessions *Sessions) error {
 	query := `
-		INSERT INTO study_sessions (title, description, subject, start_date, end_date, is_completed)
-		VALUES ($1, $2, $3, $4, $5, $6)
-		RETURNING session_id, created_at`
+    INSERT INTO study_sessions (title, description, subject, start_date, end_date, is_completed, user_id)
+    VALUES ($1, $2, $3, $4, $5, $6, $7)
+    RETURNING session_id, created_at`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
@@ -53,20 +56,22 @@ func (m *SessionsModel) Insert(sessions *Sessions) error {
 		sessions.Start_date,
 		sessions.End_date,
 		sessions.Is_completed,
+		sessions.User_id,
 	).Scan(&sessions.Session_id, &sessions.Created_at)
 }
 
 // Retrieve list of all session entries from the database
-func (m *SessionsModel) SessionList() ([]*Sessions, error) {
+func (m *SessionsModel) SessionList(userID int64) ([]*Sessions, error) {
 	query := `
-        SELECT session_id, title, description, subject, start_date, end_date, is_completed
-        FROM study_sessions
-        ORDER BY created_at DESC`
+    SELECT session_id, title, description, subject, start_date, end_date, is_completed, user_id, created_at
+    FROM study_sessions
+    WHERE user_id = $1
+    ORDER BY created_at DESC`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	rows, err := m.DB.QueryContext(ctx, query)
+	rows, err := m.DB.QueryContext(ctx, query, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -76,7 +81,7 @@ func (m *SessionsModel) SessionList() ([]*Sessions, error) {
 
 	for rows.Next() {
 		s := &Sessions{}
-		err := rows.Scan(&s.Session_id, &s.Title, &s.Description, &s.Subject, &s.Start_date, &s.End_date, &s.Is_completed)
+		err := rows.Scan(&s.Session_id, &s.Title, &s.Description, &s.Subject, &s.Start_date, &s.End_date, &s.Is_completed, &s.User_id, &s.Created_at)
 		if err != nil {
 			return nil, err
 		}
@@ -91,25 +96,40 @@ func (m *SessionsModel) SessionList() ([]*Sessions, error) {
 }
 
 // DeleteSession removes a session entry from the database using its ID
-func (m *SessionsModel) DeleteSession(sessionID int64) error {
+func (m *SessionsModel) DeleteSession(sessionID int64, userID int64) error {
 	query := `
-	DELETE FROM study_sessions WHERE session_id = $1`
+    DELETE FROM study_sessions WHERE session_id = $1 AND user_id = $2`
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	_, err := m.DB.ExecContext(ctx, query, sessionID)
-	return err
+	result, err := m.DB.ExecContext(ctx, query, sessionID, userID)
+	if err != nil {
+		return err
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rowsAffected == 0 {
+		return sql.ErrNoRows
+	}
+
+	return nil
 }
 
 // Get the session info based on the session
 func (m *SessionsModel) GetSessionByID(id int64) (*Sessions, error) {
 	stmt := `
-	SELECT session_id, title, description, subject, start_date, end_date, is_completed FROM study_sessions WHERE session_id = $1`
+    SELECT session_id, title, description, subject, start_date, end_date, is_completed, user_id, created_at
+    FROM study_sessions
+    WHERE session_id = $1`
 	row := m.DB.QueryRow(stmt, id)
 
 	var s Sessions
-	err := row.Scan(&s.Session_id, &s.Title, &s.Description, &s.Subject, &s.Start_date, &s.End_date, &s.Is_completed)
+	err := row.Scan(&s.Session_id, &s.Title, &s.Description, &s.Subject, &s.Start_date, &s.End_date, &s.Is_completed, &s.User_id, &s.Created_at)
 	if err != nil {
 		return nil, err
 	}
